Added cone volley & bullets
The end goal is to create a versatile system that can create interesting attack patterns. To do so, I decoupled the fire logic and the aiming logic.

volley is the base class that acts as an interface for all bullet spread shapes. cone_volley is the first of the bunch, which fires bullets in an evenly spread cone.

eventually I will add a separate class that decide the aim direction, which can be mixed and matched with the volleys. e.g. fixed angle, shoot at player, shoot randomly, etc.